ASG6: Split out EC2Deployment.cs;
