+++ Editor/Resources/EditorWindow/Pages/EC2Page.cs
@@ -14,8 +14,7 @@ namespace Editor.Resources.EditorWindow.Pages
     {
         private readonly VisualElement _container;
         private string _fleetName;
-        private readonly FleetParameters _parameters;
-        private DeploymentSettings _model;
+        private readonly DeploymentSettings _model;
         private readonly Button _deployButton;
         private readonly Button _redeployButton;
         private readonly Button _deleteButton;
@@ -29,7 +28,7 @@ namespace Editor.Resources.EditorWindow.Pages
             _model = DeploymentSettingsFactory.Create();
             _model.Restore();
             _model.Refresh();
-            _parameters = new FleetParameters
+            var parameters = new FleetParameters
             {
                 FleetName = _model.FleetName ?? $"{Application.productName}-ManagedFleet",
                 LaunchParameters = _model.LaunchParameters ?? $"",
@@ -47,25 +46,39 @@ namespace Editor.Resources.EditorWindow.Pages
             container.Add(uxml);
             ApplyText();
 
+            var ec2Deployment = new EC2Deployment(_model, parameters);
+
             _fleetTypeInput =
                 new FleetTypeInput(container, FleetTypeInput.ScenarioIndexMap[_model.ScenarioIndex], true);
             _fleetTypeInput.SetEnabled(true);
             _fleetTypeInput.OnValueChanged += value => { Debug.Log($"Fleet type changed to {value}"); };
 
             container.Q<Foldout>("EC2ParametersSection").text = $"{Application.productName} parameters";
-            _fleetParamsInput = new FleetParametersInput(container, _parameters);
+            _fleetParamsInput = new FleetParametersInput(container, parameters);
             _fleetParamsInput.OnValueChanged += param =>
             {
-                UpdateModelFromParameters();
+                ec2Deployment.UpdateModelFromParameters();
                 UpdateGUI();
             };
 
             _deployButton = container.Q<Button>("EC2CreateStackButton");
-            _deployButton.RegisterCallback<ClickEvent>(_ => StartDeployment());
+            _deployButton.RegisterCallback<ClickEvent>(_ =>
+            {
+                ec2Deployment.StartDeployment();
+                UpdateGUI();
+            });
             _redeployButton = container.Q<Button>("EC2RedeployStackButton");
-            _redeployButton.RegisterCallback<ClickEvent>(_ => StartDeployment());
+            _redeployButton.RegisterCallback<ClickEvent>(_ =>
+            {
+                ec2Deployment.StartDeployment();
+                UpdateGUI();
+            });
             _deleteButton = container.Q<Button>("EC2DeleteStackButton");
-            _deleteButton.RegisterCallback<ClickEvent>(_ => DeleteDeployment());
+            _deleteButton.RegisterCallback<ClickEvent>(async _ =>
+            {
+                await ec2Deployment.DeleteDeployment();
+                UpdateGUI();
+            });
             _launchClientButton = container.Q<Button>("EC2LaunchClientButton");
             _launchClientButton.RegisterCallback<ClickEvent>(_ => EditorApplication.EnterPlaymode());
 
@@ -85,52 +98,7 @@ namespace Editor.Resources.EditorWindow.Pages
             _fleetParamsInput.SetEnabled(_model.CanDeploy);
             _container.Q<Label>("EC2DeploymentStatusLabel").text = _model.CurrentStackInfo.StackStatus;
         }
-
-        private void UpdateModelFromParameters()
-        {
-            _model.FleetName = _parameters.FleetName;
-            _model.BuildName = _parameters.BuildName;
-            _model.LaunchParameters = _parameters.LaunchParameters;
-            _model.BuildFolderPath = _parameters.GameServerFolder;
-            _model.BuildFilePath = _parameters.GameServerFile;
-            _model.BuildOperatingSystem = _parameters.OperatingSystem;
-        }
-
-        private void StartDeployment()
-        {
-            UpdateModelFromParameters();
-            _model.GameName = Application.productName.Substring(0, 12);
-            if (!_model.CanDeploy) return;
-            _model.Save();
-            _model.StartDeployment(ConfirmChanges).ContinueWith(task =>
-            {
-                if (task.IsFaulted)
-                {
-                    Debug.LogException(task.Exception);
-                }
-
-                UpdateGUI();
-            });
-            UpdateGUI();
-        }
-
-        private async Task DeleteDeployment()
-        {
-            var waiter = new Waiter(CoreApi.SharedInstance);
-            waiter.InfoUpdated += () => { _model.RefreshCurrentStackInfo(); };
-            await _model.DeleteDeployment();
-            UpdateGUI();
-            await waiter.WaitUntilDone(_model);
-        }
-
-        private Task<bool> ConfirmChanges(ConfirmChangesRequest request)
-        {
-            var stackUpdateModelFactory = new StackUpdateModelFactory(new ChangeSetUrlFormatter());
-
-            StackUpdateDialog dialog = UnityEditor.EditorWindow.GetWindow<StackUpdateDialog>();
-            return dialog.SetUp(stackUpdateModelFactory.Create(request));
-        }
-
+        
         private void ApplyText()
         {
             var l = new ElementLocalizer(_container);
